refactor cephalon agent utilities into shared modules
